requirements.txt
ultralytics
python-dotenv
roboflow
opencv-python
